Updating: - Microting.eForm to 3.0.319 - Microting.eFormApi.BasePn to 1.1.178 - Microting.ItemsPlanningBase to 1.1.48 - Microsoft.NET.Test.Sdk to 16.3.0

+++ eFormAPI/Plugins/ItemsPlanning.Pn/ItemsPlanning.Pn/Handlers/eFormCaseUpdatedHandler.cs
@@ -31,7 +31,7 @@ namespace ItemsPlanning.Pn.Handlers
                 var caseDto = _sdkCore.CaseReadByCaseId(message.caseId);
                 var microtingUId = caseDto.MicrotingUId;
                 var microtingCheckUId = caseDto.CheckUId;
-                var theCase = _sdkCore.CaseRead(microtingUId, microtingCheckUId);
+                var theCase = _sdkCore.CaseRead((int)microtingUId, (int)microtingCheckUId);
 
                 itemCaseSite = SetFieldValue(itemCaseSite, theCase.Id);
